+++ impl/src/main/scala/quasar/impl/datasource/FailedDataSource.scala
@@ -18,6 +18,7 @@ package quasar.impl.datasource
 
 import slamdata.Predef.{Boolean, Unit}
 import quasar.api._, ResourceError._
+import quasar.api.datasource.DataSourceType
 import quasar.connector.DataSource
 import quasar.contrib.scalaz.MonadError_
