+++ run/src/main/scala/quasar/run/QuasarError.scala
@@ -17,7 +17,7 @@
 package quasar.run
 
 import slamdata.Predef.{Exception, Product, Serializable, Throwable}
-import quasar.api.DataSourceError.CreateError
+import quasar.api.datasource.DataSourceError.CreateError
 import quasar.compile.SemanticErrors
 import quasar.fs.Planner.PlannerError
 import quasar.sql.ParsingError
